+++ Mini_Auction/Persistence/BidDB.cs
@@ -1,9 +1,9 @@
 ﻿namespace Mini_Auction.Persistence
-using System.ComponentModel.DataAnnotations;
 {
 using System.ComponentModel.DataAnnotations.Schema;
+using System.ComponentModel.DataAnnotations;
 
-public class BidDB
+    public class BidDB
     {
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
@@ -15,7 +15,7 @@ public class BidDB
         [ForeignKey("BidderId")]
         public UserDB UserDB { get; set; }
 
-        public decimal Amount { get; set; }
+        public double Amount { get; set; }
 
         public DateTime BidTime { get; set; }
